chore: move api specs into tests/api/ directory
